LayoutTests/fast/scrolling/latching/scroll-snap-latching-expected.txt

Tests that scrolling in an overflow-scroll with scroll-snap doesn't fail to latch and scroll the document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
